restic: Unify code to load Index/Lock/Snapshot
